case-studies/log-window-analysis/jobcontainer/requirements.txt
cachetools==5.3.2 ; python_version >= "3.9" and python_version < "3.13"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "3.13"
charset-normalizer==3.0.1 ; python_version >= "3.9" and python_version < "3.13"
duckdb==0.9.1 ; python_version >= "3.9" and python_version < "3.13"
google-api-core==2.12.0 ; python_version >= "3.9" and python_version < "3.13"
google-auth==2.23.3 ; python_version >= "3.9" and python_version < "3.13"
google-cloud-core==2.3.3 ; python_version >= "3.9" and python_version < "3.13"
google-cloud-storage==2.12.0 ; python_version >= "3.9" and python_version < "3.13"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "3.13"
google-resumable-media==2.6.0 ; python_version >= "3.9" and python_version < "3.13"
googleapis-common-protos==1.61.0 ; python_version >= "3.9" and python_version < "3.13"
idna==3.4 ; python_version >= "3.9" and python_version < "3.13"
natsort==8.4.0 ; python_version >= "3.9" and python_version < "3.13"
numpy==1.26.1 ; python_version >= "3.9" and python_version < "3.13"
pandas==2.1.1 ; python_version >= "3.9" and python_version < "3.13"
protobuf==4.24.4 ; python_version >= "3.9" and python_version < "3.13"
pyasn1-modules==0.3.0 ; python_version >= "3.9" and python_version < "3.13"
pyasn1==0.5.0 ; python_version >= "3.9" and python_version < "3.13"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.13"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "3.13"
requests==2.31.0 ; python_version >= "3.9" and python_version < "3.13"
rsa==4.9 ; python_version >= "3.9" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.13"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "3.13"
urllib3==2.0.7 ; python_version >= "3.9" and python_version < "3.13"
